+++ src/_h5ai/server/php/inc/core/class-api.php
@@ -14,7 +14,7 @@ class Api {
     public function apply() {
 
         $action = Util::query_request_param("action");
-        $supported = ["download", "get", "login", "logout"];
+        $supported = ["download", "get", "login", "logout", "delete", "upload", "rename"];
         Util::json_fail(Util::ERR_UNSUPPORTED, "unsupported action", !in_array($action, $supported));
 
         $methodname = "on_${action}";
@@ -123,5 +123,84 @@ class Api {
 
         $_SESSION[AS_ADMIN_SESSION_KEY] = false;
         Util::json_exit(["asAdmin" => $_SESSION[AS_ADMIN_SESSION_KEY]]);
+    }
+
+
+    private function on_delete() {
+
+        Util::json_fail(1, "deletion disabled", !$this->options["delete"]["enabled"]);
+
+        $hrefs = Util::use_request_param("hrefs");
+
+        $hrefs = explode("|:|", trim($hrefs));
+        $errors = array();
+
+        foreach ($hrefs as $href) {
+
+        $d = Util::normalize_path(dirname($href), true);
+        $n = basename($href);
+
+            if ($this->app->is_managed_url($d) && !$this->app->is_hidden($n)) {
+
+                $path = $this->app->to_path($href);
+
+                if (!Util::delete_path($path, true)) {
+                    $errors[] = $href;
+                }
+            }
+        }
+
+        Util::json_fail(2, "deletion failed for some", count($errors) > 0);
+        Util::json_exit();
+    }
+
+
+    private function on_upload() {
+
+        Util::json_fail(1, "upload disabled", !$this->options["dropbox"]["enabled"]);
+
+        $href = Util::use_request_param("href");
+
+        Util::json_fail(2, "wrong HTTP method", strtolower($_SERVER["REQUEST_METHOD"]) !== "post");
+        Util::json_fail(3, "something went wrong", !array_key_exists("userfile", $_FILES));
+
+        $userfile = $_FILES["userfile"];
+
+        Util::json_fail(4, "something went wrong [" . $userfile["error"] . "]", $userfile["error"] !== 0);
+        Util::json_fail(5, "folders not supported", file_get_contents($userfile["tmp_name"]) === "null");
+
+        $upload_dir = $this->app->to_path($href);
+
+        Util::json_fail(6, "upload dir no h5ai folder or ignored", !$this->app->is_managed_url($href) || $this->app->is_hidden($upload_dir));
+
+        $dest = $upload_dir . "/" . urldecode($userfile["name"]);
+
+        Util::json_fail(7, "already exists", file_exists($dest));
+        Util::json_fail(8, "can't move uploaded file", !move_uploaded_file($userfile["tmp_name"], $dest));
+        Util::json_exit();
+    }
+
+
+    private function on_rename() {
+
+        Util::json_fail(1, "renaming disabled", !$this->options["rename"]["enabled"]);
+
+        $href = Util::use_request_param("href");
+        $name = Util::use_request_param("name");
+
+        $d = Util::normalize_path(dirname($href), true);
+        $n = basename($href);
+
+        if ($this->app->is_managed_url($d) && !$this->app->is_hidden($n)) {
+
+        $path = $this->app->to_path($href);
+        $folder = Util::normalize_path(dirname($path));
+
+        if (!rename($path, $folder . "/" . $name)) {
+        Util::json_fail(2, "renaming failed");
+        }
+        }
+
+        Util::json_exit();
     }
 }